HACK: proxy-exec: Readd verbose comments
Numerous questions and comments were added in this patch
series as it passed from hand to hand across a number of
developers.

Unforuntately this became distracting to patch reviewers
when submitting to the list, so they were dropped from
the original patches.

However, In order to pay heed to the travelers who have
left warnings before us, lets preserve the comments they
added here.

Signed-off-by: John Stultz <jstultz@google.com>